+++ backend/src/main/java/com/wellbeignatwork/backend/service/Evaluation/IntPointsAndGiftService.java
@@ -1,17 +1,17 @@
 package com.wellbeignatwork.backend.service.Evaluation;
 
 
-import com.wellbeignatwork.backend.entity.Evaluation.User;
+import com.wellbeignatwork.backend.entity.User;
 
 import java.util.List;
 
 public interface IntPointsAndGiftService {
 
      void save();
-     int CollectPoints(int idUser);
-     int UserGift(int idUser);
+     int CollectPoints(Long idUser);
+     int UserGift(Long idUser);
      long random(List<String> pp);
-     void UserBadge(int idUser);
+     void UserBadge(Long idUser);
      Iterable<User> PointRanking();
