+++ backend/src/main/java/com/wellbeignatwork/backend/service/Evaluation/IntVoteService.java
@@ -8,17 +8,17 @@ import com.wellbeignatwork.backend.entity.Evaluation.VoteIdea;
 import java.util.List;
 
 public interface IntVoteService {
-     int AddYes(VoteIdea v, int sujetId, int userId);
-     int AddNo(VoteIdea v, int sujetId, int userId);
-     VoteIdea getVote(int sujetId, int userId);
+     int AddYes(VoteIdea v, int sujetId, Long userId);
+     int AddNo(VoteIdea v, int sujetId, Long userId);
+     VoteIdea getVote(int sujetId, Long userId);
      void addSujet(Sujet sujet);
-     String VerifYourChoice(int userId, int sujetId);
+     String VerifYourChoice(Long userId, int sujetId);
      int countYes(int sujetId);
      int countNo(int sujetId);
      List<String> findNomdesUsersVoter(int sujetId);
-     void deletevoteById(int sujetId, int userId);
-     void UpdateYes(int sujetId, int userId);
-     void UpdateNo(int sujetId, int userId);
-     Boolean verificationvote(int sujetId, int userId);
+     void deletevoteById(int sujetId, Long userId);
+     void UpdateYes(int sujetId, Long userId);
+     void UpdateNo(int sujetId, Long userId);
+     Boolean verificationvote(int sujetId, Long userId);
 
 }
